Revert "feat: copy messages on hoverable icon (#1492)"
This reverts commit 3c1b4f627da46e4c9dc366a7c65a9ad603e2e3fa.
